% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclim.R
\docType{data}
\name{aboveground_testdata}
\alias{aboveground_testdata}
\title{aboveground data used to test the reclim package}
\description{
aboveground data used to test the reclim package
}
\examples{
data(aboveground_testdata)
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
Second dataframe (annual crop data), with 28 rows and 12 variables:
\describe{
  \item{year}{date vector, YYYY-mm-dd}
  \item{crop_description}{year, integer}
  \item{crop_id}{month, integer}
  \item{treat}{day, integer}
  \item{variance}{variance of the biomass distibution function, in days}
  \item{seeding}{seeding day, in day of year}
  \item{harvest}{first harves day, in day of year}
  \item{harvest2}{second harvest day (if present, for lay), in day of year}
  \item{tillage}{tillage day, in day of year}
  \item{minimum_cover}{minimum biomass on the ground all the time, kg of dry mass per ha}
  \item{total_dm_kg_ha}{aboveground biomass of first harvest, kg of dry mass per ha}
  \item{total_dm_kg_ha2}{aboveground biomass of second harvest (if present, for lay), kg of dry mass per ha}
  ...
}
}
