% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{add.years}
\alias{add.years}
\title{Extrapolates future climate}
\usage{
add.years(dataframe, new.years)
}
\arguments{
\item{dataframe}{the weather data frame}

\item{new.years}{the number of years you want to extend the weather data frame}
}
\value{
a table, same structure of \link{weather_testdata}
}
\description{
This function extrapolate future climate based on the average of the past 10 years (or less if less are presente)
averaging day by day
}
\examples{

add.years(weather_testdata, 5)

}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
