% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{fill.na}
\alias{fill.na}
\title{Fill NAs in the data with linear interpolation}
\usage{
fill.na(dataframe)
}
\arguments{
\item{dataframe}{the weather data frame}
}
\value{
a data frame,, same structure of the input data frame
}
\description{
This function fills NAs based on the average of the past 10 years (or less if less are presente)
averaging day by day
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
