% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAI.R
\name{GAI}
\alias{GAI}
\title{Internal function to calculate the GAI (Green Area Index) based on yields and agronomical data. The function basically distributes the biomass according to a gaussian funciton, cut at harvest.
It is useful for determining the effect of increased ET due to plants}
\usage{
GAI(yield, crop, year, variance, seeding, harvest, tillage, minimum_cover,
  yield2 = NULL, harvest2 = NULL)
}
\arguments{
\item{harvest2L}{#' @return}
}
\description{
Internal function to calculate the GAI (Green Area Index) based on yields and agronomical data. The function basically distributes the biomass according to a gaussian funciton, cut at harvest.
It is useful for determining the effect of increased ET due to plants
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
