% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{is.leapyear}
\alias{is.leapyear}
\title{Identifies if a year is leap or not}
\usage{
is.leapyear(year)
}
\arguments{
\item{year}{a year}
}
\description{
Well... pretty simple one, main heading is self-explanatory
}
\examples{

is.leapyear(1996)

}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
