% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{PET}
\alias{PET}
\title{Internal function to calculate the PET}
\usage{
PET(latitude, altitude, humidity, temperature, windspeed, sun, sun.mode, date)
}
\arguments{
\item{latitude}{self explanatory}

\item{altitude}{altitude (meters)}

\item{humidity}{(percent)}

\item{temperature}{(C)}

\item{windspeed}{in m/s}

\item{sun}{radiation, (MJ  m^-2 day-1) sunlight (0 to 1) or cloudiness (0 to 1)}

\item{sun.mode}{how sun energy is calculated (Rsolar = radiation energy, sunlight or cloudiness)}

\item{date}{a date vector}
}
\value{
A data frame of 2 variables: \code{date},\code{PET}.
}
\description{
the function is an implementation of the FAO 56 standard for evapotranspiration calculation
}
\references{
Allen, R. G., Pereira, L. S., Raes, D. & Smith, M. Crop Evapotranspiration - Guidelines for Computing Crop Water Requirements - FAO Irrigation and Drainage Paper 56. 15.
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
