% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_crop.R
\name{re_temperature}
\alias{re_temperature}
\title{internal function for determining the dependence of decomposition over soil temperature}
\usage{
re_temperature(soilT)
}
\arguments{
\item{soilT}{soil temperature daily values}
}
\value{
a vector with the daily water reduction values
}
\description{
internal function for determining the dependence of decomposition over soil temperature
}
\details{
in case of NAs in the inputs, the function attempts to fill them with a Stineman interpolation
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
