% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_crop.R
\name{re_water_moy}
\alias{re_water_moy}
\title{internal function for determining the dependence of decompositono over soil moisture}
\usage{
re_water_moy(twilt, tfield, water, porosity, L)
}
\arguments{
\item{twilt}{wilting point (0 to 1)}

\item{tfield}{field capacity (0 to 1)}

\item{water}{water balance (in mm)}

\item{porosity}{soil porosity /0 to 1)}

\item{L}{soil depth (in mm)}
}
\value{
a vector with the daily water reduction values
}
\description{
internal function for determining the dependence of decompositono over soil moisture
}
\references{
Moyano FE, Manzoni S, Chenu C. Responses of soil heterotrophic respiration to moisture availability: An exploration of processes and models. Soil Biol Biochem. Elsevier Ltd; 2013;59: 72–85. doi:10.1016/j.soilbio.2013.01.002
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
