% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclim.R
\name{reclim_annual}
\alias{reclim_annual}
\title{Just averaging the daily output from reclim to annual values}
\usage{
reclim_annual(results_daily)
}
\arguments{
\item{results_daily}{table with the first element of the list output from reclim}
}
\value{
a table
}
\description{
Just averaging the daily output from reclim to annual values
}
\examples{

reclim_annual(reclim_out$results_daily)

}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
