% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclim.R
\name{reclim_diagnostic_plots}
\alias{reclim_diagnostic_plots}
\title{Just averaging the daily output from reclim to annual values}
\usage{
reclim_diagnostic_plots(results_list, output_to_file = F)
}
\arguments{
\item{results_list}{list with the output from reclim}

\item{output_to_file}{TRUE/FALSE parameter to write the output on a file. if FALSE (default) the output goes on screen}
}
\value{
a series of plots, either on screen or on file
}
\description{
Just averaging the daily output from reclim to annual values
}
\examples{

reclim_annual(reclim_out$results_daily)

}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
