% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_crop.R
\encoding{UTF-8}
\name{soiltemp}
\alias{soiltemp}
\title{internal function for determining the soil temperature from the air temperature}
\usage{
soiltemp(L, GAI, date, temperature, LAI = NULL)
}
\arguments{
\item{L}{soil depth (mm)}

\item{GAI}{green area index daily values}

\item{date}{date vector (daily steps)}

\item{temperature}{air temperature (°C)}

\item{LAI}{(optional) LAI. If not present LAI is calculated only according to LAI=0.8*GAI, otherwise LAI is used directly}
}
\value{
a vector with the daily soil temperature values
}
\description{
internal function for determining the soil temperature from the air temperature
}
\details{
The function calculates first the surface temperature. If the temperature is below zero:
\deqn{T_{surface_i}=0.20 \cdot T}
And if the temperature is above zero
\deqn{T_{surface_i}=T_i \cdot (0.95+0.05 \cdot exp(-0.4 \cdot (LAI_i-3))}
And then calculates the soil temperature according to:
\deqn{T_{soil_{i+1}}=T_{soil_i} + (T_{surface_i} - T_{soil_i}) \cdot 0.24 \cdot e^{(-Z_{depth} \cdot 0.017)} \cdot exp(-0.15 \cdot GAI_i)}
And where \deqn{Z_{depth}=\frac{L}{20}}
The LAI is calculated as \dfunc{LAI= 0.8 \cdot GAI}
}
\references{
Kätterer, T., and O. Andrén. 2009. “Predicting Daily Soil Temperature Profiles in Arable Soils in Cold Temperate Regions from Air Temperature and Leaf Area Index.” Acta Agriculturae Scandinavica, Section B - Plant Soil Science 59 (1): 77–86. https://doi.org/10.1080/09064710801920321.
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
