% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclim.R
\docType{data}
\name{test_datasets}
\alias{test_datasets}
\alias{reclim}
\title{just some random test data used to test the reclim package}
\format{two data frames.
First dataframe, weather, with 5114 rows and 9 variables:
\describe{
  \item{date}{date vector, YYYY-mm-dd}
  \item{year}{year, integer}
  \item{month}{month, integer}
  \item{day}{day, integer}
  \item{air_temp_deg_C}{mean daily air temperature, in cat('\u00B0')C}
  \item{precipitation_mm}{cumulated daily precipitation, in mm}
  \item{windspeed_kmh}{mean daily wind speed, in km/h}
  \item{humidity_percent}{mean daily air humidity, in \%}
  \item{Rsolar}{solar radiation, in this case in Langleys. Can be also cloudiness, in hours of sun per day}
  ...
}
Second dataframe, weather, with 28 rows and 12 variables:
\describe{
  \item{year}{date vector, YYYY-mm-dd}
  \item{crop_description}{year, integer}
  \item{crop_id}{month, integer}
  \item{treat}{day, integer}
  \item{variance}{variance of the biomass distibution function, in days}
  \item{seeding}{seeding day, in day of year}
  \item{harvest}{first harves day, in day of year}
  \item{harvest2}{second harvest day (if present, for lay), in day of year}
  \item{tillage}{tillage day, in day of year}
  \item{minimum_cover}{minimum biomass on the ground all the time, kg of dry mass per ha}
  \item{total_dm_kg_ha}{aboveground biomass of first harvest, kg of dry mass per ha}
  \item{total_dm_kg_ha2}{aboveground biomass of second harvest (if present, for lay), kg of dry mass per ha}
  ...
}}
\usage{
reclim(weather, aboveground, sun.mode, latitude, altitude, depth,
  sand = NULL, clay = NULL, ave_SOC = NULL, porosity = NULL,
  wilting_point = NULL, field_capacity = NULL)
}
\description{
Data are fictional, generated from real measured data but with a great deal of noise added
}
\examples{
data(aboveground_testdata)
data(weather_testdata)

}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
