% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_crop.R
\name{waterbalance}
\alias{waterbalance}
\title{waterbalance
Internal function for the water balance model}
\usage{
waterbalance(twilt, tfield, precipitation, GAI, date, ET0, L)
}
\arguments{
\item{twilt}{wilting point (0 to 1)}

\item{tfield}{field capacity (0 to 1)}

\item{precipitation}{daily precipitations (mm)}

\item{GAI}{gren area index daily values}

\item{date}{date vector}

\item{ET0}{Evapotranspiration (calculated based on PET and GAI)}

\item{L}{soil depth (mm)}
}
\value{
a data frame with water balance and date (days)
}
\description{
waterbalance
Internal function for the water balance model
}
\examples{


}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
