% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclim.R
\docType{data}
\name{weather_testdata}
\alias{weather_testdata}
\title{weather data used to test the reclim package}
\format{
two data frames.
Data are fictional, generated from real measured data but with a great deal of noise added.
First dataframe (daily climatic values), with 5114 rows and 9 variables:
\describe{
  \item{date}{date vector, YYYY-mm-dd}
  \item{year}{year, integer}
  \item{month}{month, integer}
  \item{day}{day, integer}
  \item{air_temp_deg_C}{mean daily air temperature (°C)}
  \item{precipitation_mm}{cumulated daily precipitation, in mm}
  \item{windspeed_kmh}{mean daily wind speed, in km/h}
  \item{humidity_percent}{mean daily air humidity, in \%}
  \item{Rsolar}{solar radiation, in this case in Langleys. Can be also cloudiness, in hours of sun per day}
  ...
}
}
\description{
weather data used to test the reclim package
}
\examples{
data(weather_testdata)
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
