% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_crop.R
\name{WP}
\alias{WP}
\title{WP
Internal function for determining the soil wilting point}
\usage{
WP(sand, clay, SOC)
}
\arguments{
\item{sand}{sand %}

\item{clay}{clay % (optional)}

\item{SOC}{dSOC %}
}
\value{
a single numerical value with the soil wilting point
}
\description{
uses  Toth B, Weynants M, Nemes A, Mak? A, Bilas G, T?th G. New generation of hydraulic pedotransfer functions for Europe. Eur J Soil Sci. 2015;66: 226-238. doi:10.1111/ejss.12192
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
