% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_crop.R
\name{poros}
\alias{poros}
\title{Internal function for determining the soil porosity from texture.}
\usage{
poros(sand, clay, SOC)
}
\arguments{
\item{sand}{sand %}

\item{clay}{clay % (optional)}

\item{SOC}{dSOC %}
}
\value{
a single numerical value with the soil porosity
}
\description{
If clay is present the function uses  Toth B, Weynants M, Nemes A, Mako A, Bilas G, Toth G. New generation of hydraulic pedotransfer functions for Europe. Eur J Soil Sci. 2015;66: 226-238
otherwise  Katterer T, Andren O, Jansson P-E. Pedotransfer functions for estimating plant available water and bulk density in Swedish agricultural soils. Acta Agric Scand Sect B - Plant Soil Sci. 2006;56: 263–276
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
