% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_crop.R
\name{re_temperature}
\alias{re_temperature}
\title{re_temperature
internal function for determining the dependence of decompositono over soil temperature, sources from}
\usage{
re_temperature(soilT)
}
\arguments{
\item{soilT}{soil temperature daily values}
}
\value{
a vector with the daily water reduction values
}
\description{
re_temperature
internal function for determining the dependence of decompositono over soil temperature, sources from
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
