% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclim.R
\name{reclim_annual}
\alias{reclim_annual}
\title{function to aggregate the calculated values by year}
\usage{
reclim_annual(results_daily)
}
\arguments{
\item{results_daily}{output from \link{reclim}}
}
\value{
a table, same structure of the output from \link{reclim}
}
\description{
This function aggregates the daily results from the function \link{reclim} in annual time steps
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
