% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_crop.R
\name{WP}
\alias{WP}
\title{Internal function for determining the soil wilting point}
\usage{
WP(sand, clay, SOC)
}
\arguments{
\item{sand}{sand content \%}

\item{clay}{clay content \% (optional)}

\item{SOC}{SOC content\%}
}
\value{
a single numerical value with the soil wilting point
}
\description{
Internal function for determining the soil wilting point
}
\references{
Tóth, B., M. Weynants, A. Nemes, A. Makó, G. Bilas, and G. Tóth. 2015. “New Generation of Hydraulic Pedotransfer Functions for Europe: New Hydraulic Pedotransfer Functions for Europe.” European Journal of Soil Science 66 (1): 226–38. https://doi.org/10.1111/ejss.12192.
}
\author{
Lorenzo Menichetti \email{ilmenichetti@gmail.com}
}
